+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+0.3.0 / 2020-03-06
+==================
+
+  * feat: All prometheus_alert_group output and vars
+
 0.2.0 / 2020-02-26
 ==================